refactor: clean up logger constants and remove artifacts folder
- Replace KNOWLEDGE and SKILLS logger types with WORKSPACE
- Delete artifacts folder (was dead code from unshipped Knowledge feature)
- Move types to their proper locations:
  - LineRange → line-utils.ts
  - SearchMode → search/search-engine.ts
  - ContentSource, BaseSearchResult, etc → skills/types.ts
- Remove unnecessary re-exports
