+++ packages/core/src/logger/constants.ts
@@ -17,8 +17,7 @@ export const RegisteredLogger = {
   MCP_SERVER: 'MCP_SERVER',
   SERVER_CACHE: 'SERVER_CACHE',
   SERVER: 'SERVER',
-  KNOWLEDGE: 'KNOWLEDGE',
-  SKILLS: 'SKILLS',
+  WORKSPACE: 'WORKSPACE',
 } as const;
 
 export type RegisteredLogger = (typeof RegisteredLogger)[keyof typeof RegisteredLogger];
