+++ packages/core/src/workspace/search/search-engine.ts
@@ -5,14 +5,16 @@
  * semantic (vector), and combined hybrid search across indexed content.
  */
 
-import type { SearchMode, LineRange } from '../../artifacts';
 import type { MastraVector, VectorFilter } from '../../vector';
+import type { LineRange } from '../line-utils';
 
 import { BM25Index, tokenize, findLineRange } from './bm25';
 import type { BM25Config, TokenizeOptions } from './bm25';
 
-// Re-export SearchMode for consumers
-export type { SearchMode };
+/**
+ * Search mode options
+ */
+export type SearchMode = 'vector' | 'bm25' | 'hybrid';
 
 // =============================================================================
 // Types
