+++ packages/core/src/workspace/types.ts
@@ -1,11 +1,7 @@
 /**
  * Workspace Types
  *
- * Shared types for the workspace module to avoid circular dependencies.
+ * Shared types for the workspace module.
  */
-
-// =============================================================================
-// Status Types
-// =============================================================================
 
 export type WorkspaceStatus = 'pending' | 'initializing' | 'ready' | 'paused' | 'error' | 'destroying' | 'destroyed';
